Remove default values for rollingUpdate
RollingUpdate is an optional property only used when we set the strategy type as "RollingUpdate". If we set the type as "Recreate", this property breaks the deployments wit the following message:

Deployment.apps "<app>" is invalid: spec.strategy.rollingUpdate: Forbidden: may not be specified when strategy `type` is 'Recreate'

It gets more problematic when we use stakater/apps Helm Chart as a subchart for our applications. Helm has a known bug that doesn't allow override values from the subchart. So, if we need to deploy an application with the Recreate strategy, it is not possible as the deployment will always break with the error mentioned above.